+++ CHANGELOG.md
@@ -16,6 +16,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 #### :rocket: New Feature
 
 * Now `deprecate` returns `DeprecatedFn` `core/functools/deprecation`
+* Now `notImplement` returns `NotImplementedFn` `core/functools/not-implemented`
 
 ## v3.24.2 (2020-09-02)